return datagram size from netutils read function to allow variable length datagram reading

+++ src/net_utils/net_utils.h
@@ -87,7 +87,7 @@ uint16_t ntohs_util(uint16_t input);
  */
 int connect_wait(int sockno, struct sockaddr *addr, size_t addrlen, struct timeval *timeout);
 
-int read_poll(int sockno, int * cancel_flag, void * buf, size_t buf_size, struct timeval *timeout);
+size_t read_poll(int sockno, int * cancel_flag, void * buf, size_t buf_size, struct timeval *timeout);
 
 static const struct {
     uint64_t (*htonll)(uint64_t input);
@@ -104,7 +104,7 @@ static const struct {
 
     int (*connect)(int sockno, struct sockaddr *addr, size_t addrlen, struct timeval *timeout);
 
-    int (*read)(int sockno, int * cancel_flag, void * buf, size_t buf_size, struct timeval *timeout);
+    size_t (*read)(int sockno, int * cancel_flag, void * buf, size_t buf_size, struct timeval *timeout);
 } net_utils = {
         htonll_util,
         htonl_util,
